requirements.txt
docker==6.1.3
python-dateutil==2.8.2
pyelftools==0.30
